python-service/requirements.txt
opencv-python>=4.8.1
numpy>=1.26.0
pytesseract>=0.3.10
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.1.0
ultralytics>=8.0.0
torch>=2.1.0
torchvision>=0.16.0

# Optional: For Plate Recognizer API
# plate-recognizer==1.0.0
